refactored graphs to work on demos & train/test, async
