+++ src/mil_robogym/mil_robogym/data_collection/utils.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import os
 import re
+from collections.abc import Mapping, Sequence
 from pathlib import Path
 
 SOURCE_PROJECTS_DIR_ENV = "MIL_ROBOGYM_SOURCE_PROJECTS_DIR"
@@ -39,6 +40,31 @@ def flatten_value(value: object, prefix: str, out: dict[str, object]) -> None:
         out[prefix] = value
     else:
         out["value"] = value
+
+
+def extract_selected_state_features(
+    state: Mapping[str, object],
+    feature_names: Sequence[str],
+) -> dict[str, object]:
+    """
+    Flatten a state snapshot and return only the requested feature names.
+
+    The returned mapping preserves the order given by ``feature_names``.
+    """
+    flattened_state: dict[str, object] = {}
+
+    for topic, message in state.items():
+        temp: dict[str, object] = {}
+        flatten_value(message, "", temp)
+
+        for key, value in temp.items():
+            flattened_state[f"{topic}:{key}"] = value
+
+    return {
+        feature_name: flattened_state[feature_name]
+        for feature_name in feature_names
+        if feature_name in flattened_state
+    }
 
 
 def resolve_source_package_dir(package_name: str = "mil_robogym") -> Path:
